Revert "Big refactor to send the information after install to the app"
This reverts commit 14a936f459d6df1807decd944df4479088dacb3b.
