--- server/appmodel/config.go
@@ -1,8 +1,0 @@
-// Copyright (c) 2019-present Mattermost, Inc. All Rights Reserved.
-// See License for license information.
-
-package appmodel
-
-type MattermostConfig struct {
-	SiteURL string
-}
