+++ server/apps/get_app.go
@@ -3,13 +3,10 @@
 
 package apps
 
-import (
-	"github.com/mattermost/mattermost-plugin-apps/server/appmodel"
-	"github.com/mattermost/mattermost-plugin-apps/server/utils"
-)
+import "github.com/mattermost/mattermost-plugin-apps/server/utils"
 
 // <><> TODO remove mock, implement for real
-func (r *registry) GetApp(appID appmodel.AppID) (*appmodel.App, error) {
+func (r *registry) GetApp(appID AppID) (*App, error) {
 	app, found := r.apps[appID]
 	if !found {
 		return nil, utils.ErrNotFound
