+++ server/apps/install_app.go
@@ -8,8 +8,6 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/mattermost/mattermost-plugin-apps/server/appmodel"
-	"github.com/mattermost/mattermost-plugin-apps/server/client"
 	"github.com/mattermost/mattermost-plugin-apps/server/utils/md"
 	"github.com/mattermost/mattermost-server/v5/model"
 )
@@ -17,14 +15,14 @@ import (
 type InInstallApp struct {
 	ActingMattermostUserID string
 	NoUserConsentForOAuth2 bool
-	Manifest               *appmodel.Manifest
+	Manifest               *Manifest
 	Secret                 string
 	SessionToken           string
 }
 
 type OutInstallApp struct {
 	md.MD
-	App *appmodel.App
+	App *App
 }
 
 type InstallRequest struct {
@@ -49,7 +47,7 @@ func (r *registry) InstallApp(in *InInstallApp) (*OutInstallApp, error) {
 		return nil, err
 	}
 
-	app := &appmodel.App{
+	app := &App{
 		Manifest:               in.Manifest,
 		GrantedPermissions:     in.Manifest.RequestedPermissions,
 		NoUserConsentForOAuth2: in.NoUserConsentForOAuth2,
@@ -61,22 +59,18 @@ func (r *registry) InstallApp(in *InInstallApp) (*OutInstallApp, error) {
 	}
 	r.apps[in.Manifest.AppID] = app
 
-	c, err := client.New(in.ActingMattermostUserID, app)
-	if err != nil {
-		return nil, err
-	}
-
-	c.InstallComplete()
+	extraInfo := fmt.Sprintf(`Bot token: %s
+OAuth Client ID: %s
+OAuth Client Secret: %s`, token.Token, oAuthApp.Id, oAuthApp.ClientSecret)
 
 	out := &OutInstallApp{
-		MD:  md.Markdownf("Installed %s (%s)", in.Manifest.DisplayName, in.Manifest.AppID),
+		MD:  md.Markdownf("Installed %s (%s)\n%s", in.Manifest.DisplayName, in.Manifest.AppID, extraInfo),
 		App: app,
 	}
-
 	return out, nil
 }
 
-func (r *registry) createBot(manifest *appmodel.Manifest, sessionToken string) (*model.Bot, *model.UserAccessToken, error) {
+func (r *registry) createBot(manifest *Manifest, sessionToken string) (*model.Bot, *model.UserAccessToken, error) {
 	client := model.NewAPIv4Client(r.configurator.GetConfig().MattermostSiteURL)
 	client.SetToken(sessionToken)
 	bot := &model.Bot{
@@ -126,7 +120,7 @@ func (r *registry) createBot(manifest *appmodel.Manifest, sessionToken string) (
 	return fullBot, token, nil
 }
 
-func (r *registry) createOAuthApp(userID string, sessionToken string, manifest *appmodel.Manifest) (*model.OAuthApp, error) {
+func (r *registry) createOAuthApp(userID string, sessionToken string, manifest *Manifest) (*model.OAuthApp, error) {
 	// For the POC this should work, but for the final product I would opt for a RPC method to register the App
 	client := model.NewAPIv4Client(r.configurator.GetConfig().MattermostSiteURL)
 	app := model.OAuthApp{
