+++ server/apps/subscription.go
@@ -1,7 +1,7 @@
 // Copyright (c) 2020-present Mattermost, Inc. All Rights Reserved.
 // See License for license information.
 
-package appmodel
+package apps
 
 type SubscriptionSubject string
